update pyproject, delete free_scalar
