+++ tox.ini
@@ -4,11 +4,12 @@
 # and then run "tox" from this directory.
 
 [tox]
-envlist = py39,py310
+envlist = py38,py39,py310
 isolated_build = true
 
 [gh-actions]
 python =
+    3.8: py38
     3.9: py39
     3.10: py310
